Initial commit (V-0.1.0)
This version (0.1.0) can parse the majority of data (ignores useless info such as spawn locations) from the official valve .pop files included in the resources folder. If compiled and run it will output a few stats for every pop file. mvm_manhattan.pop has 3 bots with "class" attribute instead of "Class" which causes an issue as keywords are currently case sensative. I have corrected it to "Class" in this version but later versions should be able to handle it.